pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "hasapi"
version = "0.0.1"
description = "Modern Python web framework for AI-native APIs (LLM + RAG + Streaming)"
requires-python = ">=3.10"
license = {file = "LICENSE"}
authors = [
  { name="HasAPI Team", email="team@hasapi.dev" }
]
readme = "README.md"
keywords = ["ai", "api", "fastapi", "rag", "llm", "vector db", "asgi", "websocket"]
classifiers = [
  "Programming Language :: Python :: 3",
  "Framework :: AsyncIO",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent"
]

dependencies = [
  "httpx>=0.24",
  "orjson>=3.8",
  "uvicorn[standard]>=0.20",
  "pydantic>=2.0",
  "typing-extensions>=4.5",
  "python-dotenv>=1.0",
  "websockets>=11.0"
]

[project.optional-dependencies]
ai = ["openai>=1.0", "anthropic>=0.5"]
pytorch = ["torch>=2.0"]
onnx = ["onnxruntime>=1.15"]
gguf = ["llama-cpp-python>=0.2"]
vector = ["faiss-cpu>=1.7", "numpy>=1.24"]
all = ["torch>=2.0", "onnxruntime>=1.15", "faiss-cpu>=1.7", "llama-cpp-python>=0.2", "openai>=1.0", "anthropic>=0.5", "numpy>=1.24"]

[project.urls]
Homepage = "https://github.com/hasapi/hasapi"
Documentation = "https://hasapi.dev"
Repository = "https://github.com/hasapi/hasapi"
Issues = "https://github.com/hasapi/hasapi/issues"

[project.scripts]
hasapi = "hasapi.cli:main"

[tool.setuptools.packages.find]
include = ["hasapi*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
